# docs/swagger.yaml
basePath: /
definitions:
  service.DexPairInfo:
    properties:
      dex_slug:
        type: string
      liquidity:
        type: number
      name:
        type: string
      percent_change_price_1h:
        type: number
      percent_change_price_24h:
        type: number
      price:
        type: number
    type: object
  utils.TokenInfo:
    properties:
      price:
        type: number
      symbol:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a crypto monitoring server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Crypto Monitoring API
  version: "1.0"
paths:
  /api/v1/dex/pair:
    get:
      consumes:
      - application/json
      description: Get information about a DEX pair
      parameters:
      - description: Contract Address
        in: query
        name: contract_address
        required: true
        type: string
      - description: Network Slug
        in: query
        name: network_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.DexPairInfo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Dex Pair Info
      tags:
      - dex
  /api/v1/token/price:
    get:
      consumes:
      - application/json
      description: Get price of tokens by IDs
      parameters:
      - description: Token IDs (comma separated)
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/utils.TokenInfo'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Token Price
      tags:
      - token
  /ping:
    get:
      consumes:
      - application/json
      description: Check server health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping
      tags:
      - health
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"